[refactor] Moved Package component out of [...pkg].tsx
[ghstack-poisoned]
